Stripping out dead code
