+++ js/blockActionsHandler.js
@@ -1,178 +1,5 @@
-let isEditing = false;
-let currentEditingBlockId = null;
-
-import { categoryTags, blockTypeConfig, DEFAULT_BOOK_ACCENT } from './tagConfig.js';
 import { filterManager } from './filterManager.js';
 import { appManager } from './appManager.js';
-import { overlayHandler, getOverlayTargetTab } from './overlayHandler.js';
-import { initUsesField } from './overlayHandler.js';
-import { stripHTML } from './appManager.js';
-
-export const saveEditHandler = () => {
-    console.log("✅ Save Edit Button Clicked!");
-
-    const editOverlay = document.querySelector(".edit-block-overlay");
-
-    // Resolve the correct tab — split view carries data-active-tab on the overlay
-    const activeTab = getOverlayTargetTab(editOverlay);
-
-    const titleInput = document.getElementById("title_input_edit_overlay").value.trim();
-    const textInput  = document.getElementById("block_text_edit_overlay").innerHTML.trim();
-
-    let tagsInput = document.getElementById("tags_input_edit_overlay").value
-        .split(",")
-        .map(tag => tag.trim())
-        .filter(tag => tag.length > 0)
-        .map(tag => tag.toLowerCase());
-
-    const selectedPredefinedTags = Array.from(
-        document.querySelectorAll(".edit-block-overlay .tag-button.selected")
-    ).filter(tag => !tag.closest('#character_type_tags_edit'))
-     .map(tag => tag.dataset.tag.trim().toLowerCase());
-
-    let blocks = appManager.getBlocks(activeTab);
-
-    // In split view the block id is on the overlay; otherwise use the module-level var
-    const blockId = editOverlay?.dataset?.blockId || currentEditingBlockId;
-    const blockIndex = blocks.findIndex(block => block.id === blockId);
-
-    if (blockIndex === -1) {
-        console.error(`❌ Block with ID ${blockId} not found in ${activeTab}.`);
-        return;
-    }
-
-    const currentBlockTags = blocks[blockIndex].tags.map(tag => tag.trim().toLowerCase());
-
-    const exceptionTabs = ["tab3", "tab5", "tab6", "tab7", "tab9"];
-    if (exceptionTabs.includes(activeTab)) {
-        const dynamicTagsContainer = document.getElementById("dynamic_overlay_tags");
-        let existingUserDefinedTags = [];
-        if (dynamicTagsContainer) {
-            existingUserDefinedTags = Array.from(
-                dynamicTagsContainer.querySelectorAll(".tag-button.tag-user")
-            ).map(el => el.dataset.tag.trim().toLowerCase());
-        }
-        tagsInput = tagsInput.filter(tag => !existingUserDefinedTags.includes(tag));
-    }
-
-    tagsInput = tagsInput.filter(tag => !currentBlockTags.includes(tag));
-
-    const combinedTagsLowercase = [...new Set([...tagsInput, ...selectedPredefinedTags])];
-    const predefinedTagsMap = new Map(
-        Object.values(categoryTags).flatMap(cat => cat.tags).map(t => [t.toLowerCase(), t])
-    );
-    const allTags = combinedTagsLowercase.map(tag => {
-        const predefined = predefinedTagsMap.get(tag);
-        if (predefined) return predefined;
-        return tag.charAt(0).toUpperCase() + tag.slice(1).toLowerCase();
-    });
-
-    const selTypeBtnEdit = document.querySelector('#character_type_tags_edit .tag-button.selected');
-    const isTab3QuestEdit = activeTab === 'tab3' && selTypeBtnEdit?.dataset.tag === 'Quest';
-    const isTab3MapEdit   = activeTab === 'tab3' && selTypeBtnEdit?.dataset.tag === 'Map';
-    const textOptionalEdit = activeTab === "tab6" || isTab3QuestEdit || isTab3MapEdit;
-
-    if (!titleInput) {
-        alert("A title is required.");
-        return;
-    }
-    if (!textOptionalEdit && !textInput) {
-        alert("Title and text are required.");
-        return;
-    }
-
-    const tabBTConfig = blockTypeConfig[activeTab];
-    const blockTypeRequiredTabs = ['tab3', 'tab6', 'tab7', 'tab9'];
-    if (tabBTConfig && blockTypeRequiredTabs.includes(activeTab)) {
-        const selectedTypeBtn = document.querySelector('#character_type_tags_edit .tag-button.selected');
-        if (!selectedTypeBtn) {
-            alert(`Please select a block type: ${tabBTConfig.types.join(", ")}.`);
-            return;
-        }
-    }
-
-    const usesState = JSON.parse(localStorage.getItem("uses_field_edit_overlay_state") || "[]");
-
-    const propertiesInput = document.getElementById("properties_input_edit_overlay").value
-        .split(",")
-        .map(p => p.trim())
-        .filter(p => p.length > 0);
-
-const blockType = tabBTConfig
-        ? (tabBTConfig.singleSelect
-            ? (document.querySelector('#character_type_tags_edit .tag-button.selected')?.dataset.tag || null)
-            : Array.from(document.querySelectorAll('#character_type_tags_edit .tag-button.selected')).map(b => b.dataset.tag))
-        : null;
-
-    // Inventory-specific booleans (tab6 only) — carry through safety clears
-    let inventoryExtras = null;
-    if (activeTab === 'tab6') {
-        const requiresAttunement = !!document.getElementById('requires_attunement_edit')?.checked;
-        const equipable          = !!document.getElementById('equipable_edit')?.checked;
-        const attuned            = document.getElementById('attuned_btn_edit')?.dataset.on === 'true';
-        const equipped           = document.getElementById('equipped_btn_edit')?.dataset.on === 'true';
-        inventoryExtras = {
-            requiresAttunement,
-            equipable,
-            attuned:  requiresAttunement ? attuned  : false,
-            equipped: equipable          ? equipped : false
-        };
-    }
-
-    // Tab3-specific fields (location + quest state + book state)
-    let tab3Extras = null;
-    if (activeTab === 'tab3') {
-        const loc = document.getElementById('location_input_edit')?.value.trim();
-        tab3Extras = { location: loc || null };
-
-        const blockTypesArr = Array.isArray(blockType) ? blockType : (blockType ? [blockType] : []);
-        if (blockTypesArr.includes('Quest')) {
-            const selPill = document.querySelector('#status_buttons_edit .quest-status-pill.selected');
-            tab3Extras.status = selPill?.dataset.status || 'active';
-
-            const rows = document.querySelectorAll('#objectives_editor_edit .quest-objective-editor-row');
-            tab3Extras.objectives = Array.from(rows).map(row => ({
-                text: row.querySelector('.quest-objective-input')?.value.trim() || '',
-                done: row.dataset.done === 'true'
-            })).filter(o => o.text !== '');
-        }
-        if (blockTypesArr.includes('Book')) {
-            const descEdit  = document.getElementById('description_input_edit')?.value.trim() || '';
-            const selSwatch = document.querySelector('#book_accent_swatches_edit .book-accent-swatch.selected');
-            tab3Extras.description = descEdit;
-            tab3Extras.bookColor   = selSwatch?.dataset.colorId || DEFAULT_BOOK_ACCENT;
-        }
-        if (blockTypesArr.includes('Notes')) {
-            const descEdit = document.getElementById('description_input_edit')?.value.trim() || '';
-            tab3Extras.description = descEdit;
-        }
-        if (blockTypesArr.includes('Map')) {
-            const urlEdit = document.getElementById('url_input_edit')?.value.trim() || '';
-            if (!urlEdit) {
-                alert('A link URL is required for Map blocks.');
-                return;
-            }
-            tab3Extras.url = urlEdit;
-        }
-    }
-
-    appManager.saveBlock(
-        activeTab, titleInput, textInput, allTags, usesState, propertiesInput, blockType, blockId, blocks[blockIndex].timestamp, inventoryExtras, tab3Extras
-    );
-    console.log(`✅ Block updated in ${activeTab} with tags:`, allTags);
-
-    editOverlay.classList.remove("show");
-
-    // Clear panel context from overlay
-    delete editOverlay.dataset.activePanel;
-    delete editOverlay.dataset.activeTab;
-    delete editOverlay.dataset.blockId;
-
-    // ── Refresh the correct view ──────────────────────────────────────────────
-    filterManager.applyFilters(activeTab.replace('tab', ''));
-};
-
-export let selectedFilterTags = [];
 
 export const blockActionsHandler = (() => {
     let pendingDeleteBlockId = null;
@@ -374,93 +201,6 @@ export const blockActionsHandler = (() => {
                 appManager.enterNotesEdit(blockId);
                 return;
             }
-
-            isEditing = true;
-            currentEditingBlockId = blockId;
-        
-            selectedFilterTags = filterManager.getSelectedTags();
-            console.log("✅ Stored search & filter tags BEFORE editing:", selectedFilterTags);
-        
-            document.getElementById("title_input_edit_overlay").value = block.title;
-            const editBody = document.getElementById("block_text_edit_overlay");
-            editBody.innerHTML = block.text.replace(/\n/g, "<br>");
-            editBody.dispatchEvent(new Event('input'));
-                    
-            const usesFieldContainerEdit = document.getElementById("uses_field_edit_overlay");
-            if (usesFieldContainerEdit) {
-                localStorage.setItem("uses_field_edit_overlay_state", JSON.stringify(block.uses || []));
-                initUsesField(usesFieldContainerEdit, "uses_field_edit_overlay_state");
-            }            
-
-            const predefinedTags    = new Set(Object.values(categoryTags).flatMap(cat => cat.tags));
-            const userDefinedTags   = block.tags.filter(tag => !predefinedTags.has(tag));
-            const attachedPredefinedTags = block.tags.filter(tag => predefinedTags.has(tag));
-        
-            const exceptionTabs = ["tab3", "tab5", "tab6", "tab7", "tab9"];
-            if (exceptionTabs.includes(activeTab)) {
-                document.getElementById("tags_input_edit_overlay").value = "";
-            } else {
-                document.getElementById("tags_input_edit_overlay").value = userDefinedTags.join(", ");
-            }
-
-            document.getElementById("properties_input_edit_overlay").value = (block.properties || []).join(", ");
-
-            const editOverlay = document.querySelector(".edit-block-overlay");
-            if (editOverlay) {
-                editOverlay.dataset.activeTab = activeTab;
-                editOverlay.dataset.blockId   = blockId;
-            }
-
-            overlayHandler.initializeOverlayTagHandlers("predefined_tags_edit", block.tags);
-
-            // Populate block type buttons for this tab from config, then open overlay
-overlayHandler.populateBlockTypeOverlay("character_type_tags_edit");
-
-            // Inventory pill — show, hide, and hydrate from the block
-            const reqEdit = document.getElementById('requires_attunement_edit');
-            const eqEdit  = document.getElementById('equipable_edit');
-            const attBtnEdit = document.getElementById('attuned_btn_edit');
-            const eqBtnEdit  = document.getElementById('equipped_btn_edit');
-            if (activeTab === 'tab6') {
-                if (reqEdit) reqEdit.checked = !!block.requiresAttunement;
-                if (eqEdit)  {
-                    eqEdit.checked = !!block.equipable;
-                    eqEdit.dataset.userTouched = 'true';
-                }
-                if (attBtnEdit) attBtnEdit.dataset.on = block.attuned  === true ? 'true' : 'false';
-                if (eqBtnEdit)  eqBtnEdit.dataset.on  = block.equipped === true ? 'true' : 'false';
-            } else {
-                if (reqEdit) reqEdit.checked = false;
-                if (eqEdit)  { eqEdit.checked = false; eqEdit.dataset.userTouched = ''; }
-                if (attBtnEdit) attBtnEdit.dataset.on = 'false';
-                if (eqBtnEdit)  eqBtnEdit.dataset.on  = 'false';
-            }
-            overlayHandler.setupInventoryOverlayOptions("edit");
-            overlayHandler.setupQuestOverlayOptions("edit");
-            
-            setTimeout(() => {
-                // Re-apply selected state for regular tags
-                document.querySelectorAll(".edit-block-overlay .tag-button").forEach(tagBtn => {
-                    const tag = tagBtn.dataset.tag;
-                    if (tag && !tagBtn.closest('#character_type_tags_edit')) {
-                        tagBtn.classList.toggle("selected", block.tags.includes(tag));
-                    }
-                });
-
-                // Apply selected state for block type buttons (config-driven)
-                const tabBTConfig = blockTypeConfig[activeTab];
-                if (tabBTConfig) {
-                    const blockTypes = Array.isArray(block.blockType)
-                        ? block.blockType
-                        : (block.blockType ? [block.blockType] : []);
-                    document.querySelectorAll('#character_type_tags_edit .tag-button').forEach(btn => {
-                        btn.classList.toggle('selected', blockTypes.includes(btn.dataset.tag));
-                    });
-                }
-            }, 100);
-        
-            console.log("🟢 Edit Block Overlay Opened Successfully");
-            document.querySelector(".edit-block-overlay").classList.add("show");
 
         } else if (target.classList.contains("remove-button")) {
             showDeletePopup(blockId, event);
